+++ js/generateComment.js
@@ -1,0 +1,22 @@
+import getRandomNumber from './getRandomNumber.js';
+
+const generateComment = (id) => {
+  const avatars = ['img/avatar-1.svg', 'img/avatar-2.svg', 'img/avatar-3.svg', 'img/avatar-4.svg', 'img/avatar-5.svg', 'img/avatar-6.svg'];
+  const messages = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+
+  return {
+    id: id,
+    avatar: avatars[getRandomNumber(0, 5)],
+    message: messages[getRandomNumber(0, 5)],
+    name: 'Пользователь' + getRandomNumber(1, 100)
+  };
+}
+
+export default generateComment;
